moved script file
